Boston hackathon branch -> London hackathon branch

+++ ddl/postgres/tm_wz/wt_subject_microarray_calcs.sql
@@ -3,10 +3,10 @@
 --
 CREATE TABLE wt_subject_microarray_calcs (
     trial_name character varying(50),
-    probeset_id bigint,
-    mean_intensity bigint,
-    median_intensity bigint,
-    stddev_intensity bigint
+    probeset_id numeric(22,0),
+    mean_intensity numeric(18,6),
+    median_intensity numeric(18,6),
+    stddev_intensity numeric(18,6)
 );
 
 --
